+++ xxcmd/__init__.py
@@ -52,8 +52,10 @@ def main():
         help="Add extra padding between labels and commands.")
 
     parser.add_argument(
-        '-s', '--search-labels', action='store_const', const=True,
-        help="Search only labels not the shell commands themselves.")
+        '-s', '--search-all', action='store_const', const=True,
+        help="Search both labels and commands. Default is to search only "
+        "labels first, and only search in commands if searching for labels "
+        "resulted in no search results.")
 
     parser.add_argument(
         '-t', '--no-labels', action='store_const', const=True,
@@ -94,8 +96,9 @@ def main():
         manager.config.label_padding = args.label_padding
     if args.no_commands:
         manager.config.show_commands = not args.no_commands
-    if args.search_labels:
-        manager.config.search_labels_only = args.search_labels
+    if args.search_all:
+        manager.config.search_labels_only = False
+        manager.config.search_labels_first = False
 
     # Load db
     manager.load_database()